REAL USERREPO | Hotfix Issue #32 (#34)
* Preparation for 5/04

Ik heb TODO's toegevoegd en alles nagekeken en klaar om te DAL om te vormen. Also did some small fixed when I'm at it cause I'm a code knight.

Also 34 fucking TODO's, I'm loving it.

* Fixed Issue #22

Lap, Kleir

* (Hotfixes)

All Done, ProjectRepo is Up.

* Making sure DAL does not break you.

Commenting the red lines.

* Strange Pull (Saving my shit for a sec)

Saving For Now

* ProjectRepo = Fixed.

FINALLY.

* UserRepository Bijna Af

Just UserDetails now...

* We NEED this package

* New Issue Yay

* UserRepo Finished

* PlatformRepository Knalled

* IdeationRepo

* QuestionnaireRepository

* IdeationVoteRepo

* Temp commit 6/04

Gotta run

* IdeationQuestionsRepository

Forking finally

* UPDATE 7/04 (NOT YET FINISHED, FINISHED PAST 6PM)

* small Rename cause FORK confusion

* FINALLY ITS OVER

+++ Domain/UserInput/IOT_Device.cs
@@ -7,7 +7,7 @@ namespace Domain.UserInput
         // Added by NG
         // Modified by XV & NVZ & EKT & DM
         public int Id { get; set; }
-        public double LocationX { get; set; }
-        public double LocationY { get; set; }
+        public float LocationX { get; set; }
+        public float LocationY { get; set; }
     }
 }
